% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bts_geom}
\alias{bts_geom}
\title{Bottom trawl survey net height and spread}
\format{
A data frame
\describe{
     \item{SURVEY_DEFINITION_ID}{The survey definition ID key code is an integer that uniquely identifies a survey region/survey design.}
     \item{CRUISE}{This is a six-digit integer identifying the cruise number of the form: YYYY99 (where YYYY = year of the cruise; 99 = 2-digit number and is sequential; 01 denotes the first cruise that vessel made in this year, 02 is the second, etc.).}
     \item{YEAR}{Year the observation (survey) was collected.}
     \item{VESSEL_ID}{ID number of the vessel used to collect data for that haul.}
     \item{SURVEY_NAME}{Long name of the survey conducted}
     \item{VESSEL_NAME}{Name of the vessel used to collect data for that haul.}
     \item{DEPTH_M}{Bottom depth (meters).}
     \item{WIRE_LENGTH_M}{Length of wire deployed during a given haul in meters.}
     \item{GEAR}{Bottom trawl survey gear code.}
     \item{ACCESSORIES}{Type of accessories used on net.}
     \item{HAUL}{This number uniquely identifies a sampling event (haul) within a cruise. It is a sequential number, in chronological order of occurrence.}
     \item{DISTANCE_FISHED_KM}{Distance the net fished (thousands of kilometers).}
     \item{NET_WIDTH_M}{Measured or estimated distance (meters) between wingtips of the trawl.}
     \item{NET_HEIGHT_M}{Measured or estimated distance (meters) between footrope and headrope of the trawl.}
     \item{DURATION_HR}{This is the elapsed time between start and end of a haul (decimal hours).}
     \item{GEAR_NAME}{Gear name as a character vector.}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
bts_geom
}
\description{
Net height and spread data from good performance index station hauls from NOAA Fisheries Alaska Fisheries Science Center bottom trawl surveys in the eastern Bering Sea shelf, northern Bering Sea, eastern Bering Sea slope, Gulf of Alaska, and Aleutian Islands. Codes are described in the RACE Groundfish Asessment Program [species and data code manuals](https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual).
}
\keyword{datasets}
