% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flume_tank_rigging}
\alias{flume_tank_rigging}
\title{Flume tank rigging in 2026}
\format{
A data frame
\describe{
     \item{rig}{Letter code for the rigging configuration; used in flume_tank dataset.}
     \item{trawl}{Name of the trawl used in 2026: Poly Nor'eastern 2026 (PNE (2026)) or RACE Prototype 2026 (RACE (2026)).}
     \item{bridles}{Bridle configuration; standard = 180' bridles used on current surveys (as of 2026); 1.2: 'split' design with 45/45 upper bridles; 1: 'split' design with 70/20 upper bridles; 2: 'split' design with three ~90' bridles connected at the delta plate}
     \item{footrope}{EBS: Proposed EBS-type with 5" cookies; GOA: Proposed GOA-type with 5" cookies and 14" bobbins}
     \item{bridle_length_m}{Bridle length in the flume tank}
     \item{sweep_length_m}{Bridle length in the flume tank}
     \item{top_setback_links}{Number of links used for the top wing tip setback; each 2.5"}
     \item{middle_setback_links}{Number of links used for the middle wing tip setback; each 2.5"}
     \item{bottom_setback_links}{Number of links used for the bottom footrope setback; each 2.5"}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
flume_tank_rigging
}
\description{
Rigging information for flume tank experiments conducted at the Centre for Sustainable Aquatic Resources (CSAR), Marine Institute, Memorial University of Newfoundland, St. John's, Newfoundland and Labrador from January 27-30, 2026.
}
\keyword{datasets}
