% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bottom_contacts.R
\name{get_bottom_contacts}
\alias{get_bottom_contacts}
\title{Get individual bottom contacts}
\usage{
get_bottom_contacts(channel = NULL)
}
\arguments{
\item{channel}{Optional. An RODBC class ODBC connection}
}
\description{
A function extracting BCS data from RACEBASE.
}
