% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Build a mixed model}
\usage{
get_model(BCS_data = BCS_data)
}
\arguments{
\item{BCS_data}{Bottom contact dataframe output of get_bottom_data()}
}
\description{
A function.
}
