% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pings2}
\alias{get_pings2}
\title{Filter pings based on date/time}
\usage{
get_pings2(data)
}
\arguments{
\item{data}{data.frame containing events, measurements, and date_times for events.}
}
\description{
Filter pings based on date/time
}
