% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trawlmetrics.R
\name{get_trawlmetrics}
\alias{get_trawlmetrics}
\title{Get net mensuration data and summary statistics}
\usage{
get_trawlmetrics(
  survey,
  year,
  select_haul_types = c(3, 13, 20),
  select_gear_code = NULL,
  save_rds = FALSE,
  channel = NULL
)
}
\arguments{
\item{survey}{Survey names as a character vector. Options are: "EBS", "NBS", "GOA", "AI", "SLOPE", "CHUKCHI"}

\item{year}{Survey year as a numeric vector.}

\item{select_haul_types}{Numeric vector of valid haul types. Default: c(3, 13, 20)}

\item{save_rds}{Should output be written to an RDS file? Options are TRUE, FALSE, or a valid filepath to the location where an .rds file should be saved.}

\item{channel}{RODBC connection.}
}
\description{
Get net mensuration data and summary statistics
}
