% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{netmind_to_marport}
\alias{netmind_to_marport}
\title{Convert NetMind spread to Marport spread}
\usage{
netmind_to_marport(x)
}
\arguments{
\item{x}{Numeric vector of spread values}
}
\description{
Back-convert spread based on the conversion in Lauth and Kotwicki (2014).
}
\references{
Lauth, R. R., and S. Kotwicki. 2014. A calibration function for correcting mean net spread values obtained from Marport spread sensors used in conjunction with the Marport MK II receiver. AFSC Processed Rep. 2014-02, 26 p. Alaska Fish. Sci. Cent., NOAA, Natl. Mar. Fish. Serv., 7600 Sand Point Way NE, Seattle WA 98115. https://apps-afsc.fisheries.noaa.gov/Publications/ProcRpt/PR2014-02.pdf
}
