% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequentialOR.R
\name{sequentialOR}
\alias{sequentialOR}
\title{Sequential outlier rejection}
\usage{
sequentialOR(
  data,
  method = "lm",
  formula,
  n.reject = 1,
  n.stop,
  threshold.stop = NULL,
  tail = "both",
  ...
)
}
\arguments{
\item{data}{Data frame containing predictor and response variables.}

\item{method}{Character vector of length one indicating which function should be used to fit models. Options are 'lm', 'glm', 'gam', or 'ss'}

\item{formula}{Formula to be passed to the model. Formula must be valid for the selected method. Variables must included in `data` and the extraction operator ($) should not be used in the forumla.}

\item{n.reject}{Numerical vector of length one. Number of observations to be rejected in each iteration. Larger values speed the funciton up, but may impede precise detection of break points.}

\item{n.stop}{Numerical vector of length one which indicating the number of observations remaining or proportion of initial observations which should be remaining before outlier rejection stops. For example, with 400 initial observations, n.stop set to 40 or 0.1 (= 400*0.1) would stop the algorithm when 40 observations remain.}

\item{threshold.stop}{Optional. Threshold value for stopping sequentialOR as a 1L numeric vector. If provided, ends sequentialOR calculations when the maximum absolute residual is less than the threshold.stop.}

\item{tail}{Character vector of length one indicating whether to reject from the lower tail, upper tail, or both tails. Default = "both".}

\item{...}{Additional arguments passed to function calls (family, offset, etc.)}

\item{plot}{Logical vector indicating whether the function should print a plot of observations versus RMSE.}
}
\value{
The function returns a list with (1) input data frame with an additional column indicating the order in which observations were rejected, (2) a data frame containing the number of observations used for model-fitting and the associated RMSE. If argument `plot=T`, also prints a plot of observations versus RMSE given the remaining observations.
}
\description{
Iteratively fit a model to data and remove the most extreme residual. Three models are implemented here: ordinary least squares (lm), generalized linear model (glm), generalized additive model (GAM), or smooth spline (ss)
}
\references{
Kotwicki, S., M. H. Martin, and E. A. Laman. 2011. Improving area swept estimates from bottom trawl surveys. Fisheries Research 110(1):198–206.
}
\author{
Sean Rohan, updates by Caitlin Allen Akselrud
}
