% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sor_fill_missing.R
\name{sor_fill_missing}
\alias{sor_fill_missing}
\title{Fill missing height and spread}
\usage{
sor_fill_missing(
  height_paths,
  spread_paths,
  rds_dir,
  convert_marport_to_netmind = TRUE
)
}
\arguments{
\item{height_paths}{Path(S) to height data files.}

\item{spread_paths}{Path(s) to spread data files.}

\item{rds_dir}{Path to haul rds files.}

\item{convert_marport_to_netmind}{Should Marport spread measurements be converted to Netmind spread using trawlmetric::marport_to_netmind()?}
}
\value{
Reads in measurement data from _sor.rds files from rds_dir and writes corrected results to _final.rds files in rds_dir.
}
\description{
Fill missing height and spread
}
