% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sor_run.R
\name{sor_run}
\alias{sor_run}
\title{Function to run sequential outlier rejection on data}
\usage{
sor_run(
  vessel = NULL,
  cruise = NULL,
  region = NULL,
  survey = NULL,
  min_pings_for_sor = 50
)
}
\arguments{
\item{vessel}{vessel ID number as a numeric vector (e.g. 162 for Alaska Knight. Must provide rds_dir or all of region, vessel, cruise, survey.}

\item{cruise}{Cruise number as a numeric vector (e.g. 202202). Must provide rds_dir or all of region, vessel, cruise, survey.}

\item{region}{Survey region as a 1L character vector (EBS or NBS). Must provide rds_dir or all of region, vessel, cruise, survey.}

\item{survey}{Survey name prefix to use in filename (e.g. NBS_2022). Must provide rds_dir or all of region, vessel, cruise, survey.}

\item{min_pings_for_sor}{Minimum number of ping required to conduct sequential outlier rejection (default = 50).}
}
\value{
Reads in measurement data from _ping.rds files from rds_dir and writes corrected results to _sor.rds files in rds_dir.
}
\description{
Runs sequential outlier rejections on hauls that have at least 50 spread measurements.
}
