% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sor_save_results.R
\name{sor_save_results}
\alias{sor_save_results}
\title{Setup directory}
\usage{
sor_save_results(
  final_dir,
  create_user = "",
  survey,
  cruise_idnum,
  channel = NULL,
  delete_existing = TRUE
)
}
\arguments{
\item{final_dir}{Path to one or more directories containing outputs of sequential outlier rejection process. Filenames will end in 'final.rds'}

\item{create_user}{Oracle user making the update.}

\item{survey}{Survey name prefix to use in output filename (e.g. NBS_2022)}

\item{cruise_idnum}{One or more cruise ID numbers as a numeric vector (e.g. 757). Must be the same length as final_dir.}

\item{channel}{An RODBC channel. Will prompt user to get connected if NULL.}

\item{delete_existing}{Logical indicating whether to overwrite data in RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES}
}
\description{
Function to retrieve data for sequential outlier rejection from RACEBASE and split data by haul.
}
