% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawlmetrics-package.R
\docType{package}
\name{trawlmetrics-package}
\alias{trawlmetrics}
\alias{trawlmetrics-package}
\title{trawlmetrics: Trawl effort correction and gear performance measures}
\description{
Perform trawl effort correction and retrieve gear performance measures.
}
\author{
\strong{Maintainer}: Sean Rohan \email{sean.rohan@noaa.gov}

Authors:
\itemize{
  \item Caitlin Allen-Akselrud \email{caitlin.allen_akselrud@noaa.gov}
  \item Shannon Nardi \email{shannon.nardi@noaa.gov}
}

Other contributors:
\itemize{
  \item Liz Dawson \email{liz.dawson@noaa.gov} [contributor]
  \item Duane Stevenson \email{duane.tevenson@noaa.gov} [contributor]
  \item Stan Kotwicki \email{stan.kotwicki@noaa.gov} [contributor]
}

}
\keyword{internal}
