% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flume_tank}
\alias{flume_tank}
\title{Flume tank experimental data}
\format{
A data frame
\describe{
     \item{date_time_nst}{Date and time in Newfoundland Standard Time (UTC-3.5).}
     \item{trial}{Trial number.}
     \item{trawl}{Trawl model (e.g., "83-112", "PNE", "RACE").}
     \item{bridles}{Type of bridles used (e.g., "standard", "bridle 1", "bridle 2").}
     \item{footrope}{Footrope used on the trawl (e.g., "EBS" or "GOA").}
     \item{pulling_point_elevation_m}{Elevation of the pulling point on the trawl mast in meters.}
     \item{backstrap_length_m}{Backstrap length in meters.}
     \item{bridle_u_length_m}{Upper bridle length in meters.}
     \item{bridle_l_length_m}{Lower bridle length in meters.}
     \item{sweep_length_m}{Sweep length in meters.}
     \item{u_bridle_extension_m}{Upper bridle extension length in meters.}
     \item{towing_speed_kn}{Towing speed in knots.}
     \item{spread_door_m}{Door spread in meters.}
     \item{spread_u_wing_m}{Upper wing spread in meters.}
     \item{spread_l_wing_m}{Lower wing spread in meters.}
     \item{spread_mean_we_m}{Mean wing spread in meters.}
     \item{opening_wing_m}{Wing opening height in meters.}
     \item{opening_headline_m}{Headline opening height in meters.}
     \item{height_headline_m}{Headline opening height from the bottom in meters.}
     \item{bridle_tension_port_t}{Port bridle tension in metric tons.}
     \item{bridle_tension_stbd_t}{Starboard bridle tension in metric tons.}
     \item{bridle_tension_total_t}{Total bridle tension in metric tons.}
     \item{mouth_area_m2}{Mouth opening area in square meters.}
     \item{mouth_drag_kgf_m2}{Mouth drag in kilogram-force per square meter.}
     \item{bridle_angle_deg}{Bridle angle of attack in degrees.}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
flume_tank
}
\description{
Net measurement data from flume tank experiments conducted at the Centre for Sustainable Aquatic Resources (CSAR), Memorial University of Newfoundland, St. John's, Newfoundland and Labrador from January 13-17, 2025. CURRENTLY EXAMPLE DATA.
}
\keyword{datasets}
