% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bottom_data.R
\name{get_bottom_data}
\alias{get_bottom_data}
\title{Get bottom contact summary stats}
\usage{
get_bottom_data(channel = NULL, contact_dat = contact_dat)
}
\arguments{
\item{channel}{Optional. An RODBC class ODBC connection}

\item{contact_dat}{Bottom contact dataframe output by get_bottom_contacts()}
}
\description{
A function that builds a dataframe from BCS contacts.
}
